refactored files and functions
